Added tests for CacheManager
