+++ rssparser/src/test/java/com/prof/rssparser/core/CoreXMLParserFeedRuTest.kt
@@ -15,15 +15,12 @@
 *
 */
 
-package com.prof18.rssparser
+package com.prof.rssparser.core
 
 import android.os.Build
 import com.prof.rssparser.Article
 import com.prof.rssparser.Channel
-import com.prof.rssparser.Image
-import com.prof.rssparser.core.CoreXMLParser
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNull
+import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -32,10 +29,10 @@ import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [Build.VERSION_CODES.P])
-class CoreXMLParserAudioFeedTest {
+class CoreXMLParserFeedRuTest {
     private lateinit var articleList: MutableList<Article>
     private lateinit var article: Article
-    private val feedPath = "/feed-test-audio.xml"
+    private val feedPath = "/feed-test-ru.xml"
     private lateinit var channel: Channel
 
     @Before
@@ -49,33 +46,27 @@ class CoreXMLParserAudioFeedTest {
 
     @Test
     fun channelTitle_isCorrect() {
-        assertEquals(channel.title, "Stuff You Should Know")
+        assertEquals(channel.title, "Аргументы и Факты")
     }
 
     @Test
     fun channelDesc_isCorrect() {
-        assertEquals(channel.description, "If you've ever wanted to know about champagne, satanism, the Stonewall Uprising, chaos theory, LSD, El Nino, true crime and Rosa Parks, then look no further. Josh and Chuck have you covered.")
+        assertEquals(channel.description, "Аргументы и Факты: объясняем, что происходит")
     }
 
     @Test
     fun channelLink_isCorrect() {
-        assertEquals(channel.link, "https://www.howstuffworks.com")
+        assertEquals(channel.link, "http://www.aif.ru/")
     }
 
     @Test
     fun channelImage_isNull() {
-        assertEquals(channel.image,
-                Image(
-                        title = "Stuff You Should Know",
-                        url = "https://megaphone-prod.s3.amazonaws.com/podcasts/1e705dd4-2de6-11e8-b55d-9ba6ddb3f75e/image/uploads_2F1546996139536-0o3pw93d8mk-d5f1143c14a746754c55efb478c66988_2FSKSKLogo-FINAL-iHR-3000x3000.png",
-                        link = "https://www.howstuffworks.com"
-                )
-        )
+        assertNull(channel.image)
     }
 
     @Test
     fun channelLastBuild_isCorrect() {
-        assertNull(channel.lastBuildDate)
+        assertEquals(channel.lastBuildDate, "2019-05-18 10:58:46 +0300")
     }
 
     @Test
@@ -86,13 +77,13 @@ class CoreXMLParserAudioFeedTest {
     @Test
     @Throws
     fun size_isCorrect() {
-        assertEquals(articleList.size, 1)
+        assertEquals(articleList.size, 100)
     }
 
     @Test
     @Throws
     fun title_isCorrect() {
-        assertEquals(article.title, "SYSK Selects: How Gold Works")
+        assertEquals(article.title, "СМИ: собака в Таиланде спасла заживо похороненного младенца")
     }
 
     @Test
@@ -104,50 +95,51 @@ class CoreXMLParserAudioFeedTest {
     @Test
     @Throws
     fun link_isCorrect() {
-        assertNull(article.link)
+        assertEquals(article.link, "http://www.aif.ru/incidents/smi_sobaka_v_tailande_spasla_zazhivo_pohoronennogo_mladenca")
     }
 
     @Test
     @Throws
     fun pubDate_isCorrect() {
-        assertEquals(article.pubDate, "Sat, 07 Mar 2020 10:00:00 -0000")
+        assertEquals(article.pubDate, "Sat, 18 May 2019 10:52:50 +0300")
     }
 
     @Test
     @Throws
     fun description_isCorrect() {
-        assertEquals(article.description, "As of early 2013, only 161,00 metric tons of gold had been mined in the entire history of the world. Considering about 85 percent of it is recycled, there's a decent chance your jewelry may once have been part of an Incan headdress or Mycenaean face mask. Dive in to gold in this classic episode.\n" +
-                "                Learn more about your ad-choices at https://news.iheart.com/podcast-advertisers")
+        assertEquals(article.description, "15-летняя мать ребенка решила избавиться от него, побоявшись гнева родителей")
     }
 
     @Test
     @Throws
     fun content_isCorrect() {
-        assertEquals(article.content, "<p>As of early 2013, only 161,00 metric tons of gold had been mined in the entire history of the world. Considering about 85 percent of it is recycled, there's a decent chance your jewelry may once have been part of an Incan headdress or Mycenaean face mask. Dive in to gold in this classic episode.</p><p> </p> Learn more about your ad-choices at <a href=\"https://news.iheart.com/podcast-advertisers\">https://news.iheart.com/podcast-advertisers</a>")
+        assertEquals(article.content, null)
     }
 
     @Test
     @Throws
     fun image_isCorrect() {
-        assertNull(article.image)
+        assertEquals(article.image, "https://images.aif.ru/017/020/025bc5cb4cc5d0f8347bbb27f5e4d13b.jpg")
     }
 
     @Test
     @Throws
     fun categories_isCorrect() {
-        assertEquals(article.categories.size, 0)
+        assertEquals(article.categories, mutableListOf(
+                "Происшествия"
+        ))
     }
 
     @Test
     @Throws
     fun guid_isCorrect() {
-        assertEquals(article.guid, "c41042ae-5460-11e8-b38c-5f5faf9dd0c5")
+        assertEquals(article.guid, null)
     }
 
     @Test
     @Throws
     fun audio_iCorrect() {
-        assertEquals(article.audio, "https://www.podtrac.com/pts/redirect.mp3/chtbl.com/track/5899E/traffic.megaphone.fm/HSW3157672398.mp3")
+        assertNull(article.audio)
     }
 
     @Test
