+++ rssparser/src/test/java/com/prof/rssparser/core/CoreXMLParserStandardFeedTest.kt
@@ -15,12 +15,11 @@
 *
 */
 
-package com.prof18.rssparser
+package com.prof.rssparser.core
 
 import android.os.Build
 import com.prof.rssparser.Article
 import com.prof.rssparser.Channel
-import com.prof.rssparser.core.CoreXMLParser
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
